% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{sol_allometry}
\alias{sol_allometry}
\title{Apply allometric equations}
\usage{
sol_allometry(data, equation_id)
}
\arguments{
\item{data}{data.frame: input data}

\item{equation_id}{string or character: the identifier of the equation to apply. Either a single string (that equation will be applied to all rows of the data) or a character vector with length matching the number of rows of the data}
}
\value{
the input data frame, augmented with columns "allometric_property", "allometric_value", "allometric_value_lower", and "allometric_value_upper"
}
\description{
Apply allometric equations
}
\examples{
\dontrun{
x <- data.frame(LRL=c(11.3,13.9),species=c("Architeuthis dux"),
  stringsAsFactors=FALSE)
## it doesn't matter what the column names are, but we
## need to set the property types correctly
x$LRL <- sol_set_property(x$LRL,"lower rostral length")

## apply a single equation to all rows
sol_allometry(x,c("342218_ML_Roel2000"))

## apply a different equation to each row
sol_allometry(x,c("342218_ML_Roel2000","342218_ML_Clar1986"))
}
}
\seealso{
\code{\link{allometric_equations}}
}
