% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{sol_make_equation}
\alias{sol_make_equation}
\title{Create an allometric equation object}
\usage{
sol_make_equation(equation_id, taxon_name, taxon_aphia_id, equation, inputs,
  return_property, return_units, reliability, notes, reference,
  check_packaged_ids = TRUE, warn_recommended = TRUE)
}
\arguments{
\item{equation_id}{string: a unique identifier for the equation (required)}

\item{taxon_name}{string: the taxon name that the equation applies to (required)}

\item{taxon_aphia_id}{numeric: the AphiaID of the taxon that the equation applies to (recommended)}

\item{equation}{function: the equation; see details (required)}

\item{inputs}{data.frame: the inputs needed by the equation; see details (required)}

\item{return_property}{string: the name of the allometric property that the equation returns (required)}

\item{return_units}{string: the units of measurement of the allometric property that the equation returns. Must be units that are recognized by units::ud_unit (required)}

\item{reliability}{data.frame: indicators of reliability of the equation; see details (recommended)}

\item{notes}{string: any notes that users should be aware of (optional)}

\item{reference}{string: the source of the equation (recommended)}

\item{check_packaged_ids}{logical: if TRUE, check the equation_id against the package-bundled equations. A warning will be issued if there is a packaged equation with the same ID as equation_id}

\item{warn_recommended}{logical: issue a warning if "recommended" informations is not supplied?}
}
\value{
equation object
}
\description{
Create an allometric equation object
}
\examples{
library(dplyr)
eq <- sol_make_equation(equation_id="my_equation_id",
                        taxon_name="Chorismus antarcticus",
                        taxon_aphia_id=369214,
                        equation=function(L)
                           tibble(allometric_value=0.000943*(L^2.976)),
                        inputs=tibble(property="carapace length",units="mm"),
                        return_property="mass",
                        return_units="g",
                        reliability=tribble(~type,~value,
                                             "N",35,
                                             "R^2",0.976),
                        reference="Lake S et al. (2003) doi:10.3354/meps254293")

}
\seealso{
\code{\link{sol_equation}} \code{\link{sol_equations}}
}
