% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{sol_properties}
\alias{sol_properties}
\title{Properties}
\usage{
sol_properties(prop)
}
\arguments{
\item{prop}{string: if provided, return only the property matching this name}
}
\value{
data.frame
}
\description{
Properties
}
\examples{
sol_properties() ## all properties that solong knows about

}
\seealso{
\code{\link{sol_set_property}}
}
