% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sol_properties_data}
\alias{sol_properties_data}
\title{Properties relating to allometric equations.}
\format{A data frame with variables:
\describe{
  \item{property}{the property name}
  \item{units}{its units of measurement}
  \item{class_name}{the corresponding class name used internally by the solong package}
}}
\usage{
sol_properties_data
}
\description{
Properties relating to allometric equations.
}
\keyword{datasets}
