% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{sol_set_property}
\alias{sol_set_property}
\alias{sol_get_property}
\title{Set or get the property name}
\usage{
sol_set_property(x, prop, with_units, ...)

sol_get_property(x)
}
\arguments{
\item{x}{vector: data}

\item{prop}{string: property name}

\item{with_units}{string: units of measurement to use. If missing, the default units for the property will be used}

\item{...}{: extra arguments, currently ignored}
}
\value{
x with additional class set
}
\description{
Set or get the property name
}
\examples{
x <- data.frame(LRL=c(11.3,13.9),species=c("Architeuthis dux"),
  stringsAsFactors=FALSE)
## it doesn't matter what the column names are, but we
## need to set the property types correctly
x$LRL <- sol_set_property(x$LRL,"lower rostral length")

## remove the property
x$LRL <- sol_set_property(x$LRL,NULL)

}
\seealso{
\code{\link{sol_properties}}
}
