% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sol_vonbert}
\alias{sol_vonbert}
\title{Von Bertalanffy growth equation with optional propagation of uncertainty}
\usage{
sol_vonbert(
  t,
  Linf,
  k,
  t0,
  Linf_se,
  k_se,
  t0_se,
  reps = 1000L,
  ci = 0.95,
  method = "monte carlo"
)
}
\arguments{
\item{t}{numeric: prediction times}

\item{Linf}{numeric: Linf parameter estimate}

\item{k}{numeric: k parameter estimate}

\item{t0}{numeric: t0 parameter estimate}

\item{Linf_se}{numeric: Linf parameter standard error estimate}

\item{k_se}{numeric: k parameter standard error estimate}

\item{t0_se}{numeric: t0 parameter standard error estimate}

\item{reps}{integer: number of Monte-Carlo samples to draw}

\item{ci}{numeric: confidence level}

\item{method}{string: method to use for uncertainty propagation (only "monte carlo" supported at the moment)}
}
\value{
A tibble with columns \code{allometric_value}, \code{allometric_value_lower}, \code{allometric_value_upper}. If any of the standard error inputs are missing, \code{NULL}, or \code{NA}, the upper and lower estimates will be \code{NA_real_}
}
\description{
Intended for use inside of \code{sol_equation} objects.
}
\examples{

sol_vonbert_mc(0:7, Linf = 80.7, Linf_se = 0.82, k = 0.25, k_se = 0.01, t0 = -2.31, t0_se = 0.01)

}
