% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{sol_equation}
\alias{sol_equation}
\title{Allometric equations}
\usage{
sol_equation(id)
}
\arguments{
\item{id}{string: the identifier of the equation. If omitted, all equations will be returned}
}
\value{
data.frame
}
\description{
Allometric equations
}
\examples{
sol_equation()
sol_equation("342218_ML_Roel2000")

}
