% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{sol_fb_length_weight}
\alias{sol_fb_length_weight}
\title{Create allometric equation from Fishbase}
\usage{
sol_fb_length_weight(..., worms = requireNamespace("worrms", quietly = TRUE))
}
\arguments{
\item{...}{: arguments passed to rfishbase::length_weight}

\item{worms}{logical: if TRUE, and if the worrms package is installed, try and find the AphiaID for the taxon in the World Register of Marine Species}
}
\value{
equation object
}
\description{
Experimental. Requires the rfishbase package to be installed.
}
\examples{
\dontrun{
 library(dplyr)
 eq <- sol_fb_length_weight("Electrona antarctica")
 x <- tibble(SL = 10) \%>\%
   mutate(SL = sol_set_property(SL, "standard length", with_units = "cm"))
 sol_allometry(x, eq[1, ])
}

}
\seealso{
\code{\link{sol_equation}} \code{\link{sol_equations}}
}
