% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{sol_fb_length_weight}
\alias{sol_fb_length_weight}
\title{Create allometric equation from Fishbase}
\usage{
sol_fb_length_weight(...)
}
\arguments{
\item{...}{: arguments passed to rfishbase::length_weight}
}
\value{
equation object
}
\description{
Experimental.
}
\examples{
\dontrun{
 library(dplyr)
 eq <- sol_fb_length_weight("Electrona antarctica")
 sol_allometry(tibble(SL=10 \%>\% sol_set_property("standard length",with_units="cm")),eq[2,])
}

}
\seealso{
\code{\link{sol_equation}} \code{\link{sol_equations}}
}
